Catch ContentTypes errors if they're run before migrations
