+++ payments/notifications/types.py
@@ -1,0 +1,9 @@
+from django.db.models import TextChoices
+from django.utils.translation import gettext_lazy as _
+
+
+class NotificationType(TextChoices):
+    ORDER_APPROVED = (
+        "order_approved",
+        _("Order approved"),
+    )
